requirements.txt
pillow==10.2.0
ttkthemes==3.2.2
pandas==2.2.1
openpyxl==3.1.2 